test: set base stake amount to 0 before drain validator
